Added Run method and deprecated Start
This change deprectates leaky Start method and suggests
using the new Run method instead. The Run method blocks
until all goroutines exit.

The main issue with previous implementation is that it’s
not possible to wait for worker exit. This change also fixes
time.After leak on context cancellation by explicitly managing
and reusing time.Timer.

It’s also common to have a callback in Run method to signal
that the startup has completed, e.g.

	// This example simplifies things a bit but the general
	// idea is to notify on a channel about successful start
	// or error. Usually there is also another goroutine that
	// selects on error channel even after startup to handle
	// unexpected shutdowns.
	errc := make(chan error, 1)
	done := make(chan struct{}, 1)
	grp.Go(func() error {
		errc <- w.Run(ctx, func(ctx context.Context) error {
			done <- struct{}{}
			<-ctx.Done()
			return nil
		})
		return nil
	})
	select {
	case <-done:
		fmt.Println("up and running")
	case <-errc:
		fmt.Println("failed to start")
	}

That’s not neccesary with Worker and WorkerPool since all
interactions are done via Client and any startup error is
considered unrecoverable. In terms of systemd, the pattern
above corresponds to “notify” service type while we use
the “exec” type for workers. And the previous behavior is
like “forking” type without PID file, ouch.